refactor(pipeline): extract the sentiment identity into pipeline/sentiment.py
Pure move: model, sampling, pricing, frozen prompt, and parser leave
batch.py unchanged; batch.py imports them for now. ClassifierStage
(pipeline/stage.py) names a classification pass; both stages declare
an instance so the transport and eval harness can take one (#97).
